# bring back transformed documents