so I can use collection helpers as I did with $collection service.
Also added small fix to getCollectionByName('users')