@Urigo 2015-02-15
Looks very good.
But can you demonstrate the use case for the collection helpers?  (just for my personal knowledge)