@Nieziemski 2015-02-15
Sure,

I use collection helpers package: https://github.com/dburles/meteor-collection-helpers
Here's an example:
```javascript
// register helper
Meteor.users.helpers({
    getFullName: function() {
        return this.firstName + ' ' + this.lastName;
    }
});

// use in meteor world
Template.profile.helpers({
    getName: function(userId) {
        var user = Meteor.users.findOne(userId);
        return user.getFullName();
    }
})

// and angular
function profileController($scope, $collection) {
    $collection(Meteor.users).bindOne($scope, 'user', someId);
}

<div ng-controller="profileController">
    [[user.getFullName()]]
</div>
```

It breaks separation of angular and meteor code, but it's pretty handy ;)