@Urigo 2015-02-16
I'm a little bit concerned about performance...
Can you test that a little bit and see the difference?