@Nieziemski 2015-02-16
Never did it before, and I'm not sure how to do it properly. I use chrome profiler, start recording, on click create collection and look for total AngularMeteorCollection.prototype.updateCursor on list.
For collection with 100 elements not binded to DOM I've got results:

 Old code:
532.8 ms
535.7 ms
449.3 ms

New code:
628.1 ms
729.0 ms
802.0 ms

So, as expected, it is slower :/