@netanelgilad 2015-02-17
Hey @Nieziemski,
As I see it, what you actually need from the collection object is the extra methods supplied on it by the collection helper package.
Have you thought about assigning the prototype instead of getting the object from the collection? That may be good for performance but still get the effect of having the wanted methods on the collection object..
Try and see if that works for you too..