@Urigo 2015-02-18
@Nieziemski,  @netanelgilad suggestion sounds good to me.
Can you try that quickly and let me know how it works?