@Nieziemski 2015-02-19
I couldn't belive that findOne() causes such a performance decrease, so I investigated and found a potential improvement. I call digest only once after cursor update, not after each element change. Running previous tests I get results like this:

Old code (with new safeApply()):
4.8 ms
2.5 ms
10.8 ms

New code (with new safeApply()):
51.5 ms
23.6 ms
40.4 ms

Previously, when I binded this 100 elements in template page rendered in 14000 ms (!) and now it's 529.9 ms.
It shouldn't break anything (works for me perfectly), but can you take a look and check if I'm not missing something?
findOne() still is decreasing performance though...