@Urigo 2015-02-19
@Nieziemski that's really cool.

Let's do this:
* Separate your pull request to different ones - 

1. Fixing the users collection - no need for that anymore as I'm using a package for that now
2. Fixing digest - it's a big change so I want to test that separately
3. Getting the collection - test that with @netanelgilad solution, that way we might get everything with even better performance

What do you think?