@Nieziemski 2015-02-19
@netanelgilad Yes, in my case it's all about extra methods. But in general transform() can do something else with document. Running self.$$collection.transform() is another solution here. I tested it and results are similar to findOne() version. It may be more efficient if we have huge amount of documents, becouse findOne() should perform worse. In the other hand if transform() is expensive it should be less efficient, becouse complexity is linear then (and we call it second time). 
I still think that we should keep transformed documents in collection and findOne() version is better ;)