@Urigo 2015-02-19
@Nieziemski ok, let's do that.
Then we can test separately about improving performance with transform().
But please open the digest code in a separate pull request and remove and "users" code.
Then let's merge it already!