@Urigo 2015-02-19
@Nieziemski I want to release a new version today or tomorrow so looking forward for your changes