@Nieziemski 2015-02-20
@Urigo I created new pull requests. Tests fails, but i see that the reason is in previous commits. I will be back on monday, hope everything goes well.